backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
aiosqlite==0.20.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.7.1
python-dotenv==1.0.1
starlette==0.37.2
anyio==4.3.0
dnspython==2.6.1
